+++ contracts/p1/BackingManager.sol
@@ -236,6 +236,7 @@ contract BackingManagerP1 is TradingP1, IBackingManager {
         for (uint256 i = 0; i < length; ++i) {
             IAsset asset = assetRegistry.toAsset(erc20s[i]);
 
+            // Use same quantity-rounding as BasketHandler.basketsHeldBy()
             // {tok} = {BU} * {tok/BU}
             uint192 req = needed.mul(basketHandler.quantity(erc20s[i]), CEIL);
             uint192 bal = asset.bal(address(this));