+++ contracts/plugins/assets/RTokenAsset.sol
@@ -57,6 +57,8 @@ contract RTokenAsset is IAsset, VersionedAsset, IRTokenOracle {
     /// @return low {UoA/tok} The low price estimate
     /// @return high {UoA/tok} The high price estimate
     function tryPrice() external view virtual returns (uint192 low, uint192 high) {
+        // Get the BU price /w issuance premium. This means in UoA terms that the
+        // high price of a BU stays just above peg, even in the midst of a default.
         (uint192 lowBUPrice, uint192 highBUPrice) = basketHandler.price(); // {UoA/BU}
         require(lowBUPrice != 0 && highBUPrice != FIX_MAX, "invalid price");
         assert(lowBUPrice <= highBUPrice); // not obviously true just by inspection
@@ -88,7 +90,7 @@ contract RTokenAsset is IAsset, VersionedAsset, IRTokenOracle {
     }
 
     /// Should not revert
-    /// @dev See `tryPrice` caveat about possible compounding error in calculating price
+    /// @dev See `tryPrice` caveats
     /// @return {UoA/tok} The lower end of the price estimate
     /// @return {UoA/tok} The upper end of the price estimate
     function price() public view virtual returns (uint192, uint192) {