+++ test/Main.test.ts
@@ -1762,10 +1762,10 @@ describe(`MainP${IMPLEMENTATION} contract`, () => {
       it('Should not allow to set prime Basket with invalid length', async () => {
         await expect(
           basketHandler.connect(owner).setPrimeBasket([token0.address], [])
-        ).to.be.revertedWith('len mismatch')
+        ).to.be.revertedWith('invalid lengths')
         await expect(
           basketHandler.connect(owner).forceSetPrimeBasket([token0.address], [])
-        ).to.be.revertedWith('len mismatch')
+        ).to.be.revertedWith('invalid lengths')
       })
 
       it('Should not allow to set prime Basket with non-collateral tokens', async () => {
@@ -1830,10 +1830,10 @@ describe(`MainP${IMPLEMENTATION} contract`, () => {
 
       it('Should not allow to set prime Basket with an empty basket', async () => {
         await expect(basketHandler.connect(owner).setPrimeBasket([], [])).to.be.revertedWith(
-          'empty basket'
+          'invalid lengths'
         )
         await expect(basketHandler.connect(owner).forceSetPrimeBasket([], [])).to.be.revertedWith(
-          'empty basket'
+          'invalid lengths'
         )
       })
 
@@ -2065,10 +2065,10 @@ describe(`MainP${IMPLEMENTATION} contract`, () => {
       it('Should not allow to set prime Basket with invalid length', async () => {
         await expect(
           indexBH.connect(owner).setPrimeBasket([token0.address], [])
-        ).to.be.revertedWith('len mismatch')
+        ).to.be.revertedWith('invalid lengths')
         await expect(
           indexBH.connect(owner).forceSetPrimeBasket([token0.address], [])
-        ).to.be.revertedWith('len mismatch')
+        ).to.be.revertedWith('invalid lengths')
       })
 
       it('Should not allow to set prime Basket with non-collateral tokens', async () => {
@@ -2123,10 +2123,10 @@ describe(`MainP${IMPLEMENTATION} contract`, () => {
 
       it('Should not allow to set prime Basket with an empty basket', async () => {
         await expect(indexBH.connect(owner).setPrimeBasket([], [])).to.be.revertedWith(
-          'empty basket'
+          'invalid lengths'
         )
         await expect(indexBH.connect(owner).forceSetPrimeBasket([], [])).to.be.revertedWith(
-          'empty basket'
+          'invalid lengths'
         )
       })
 
@@ -3536,15 +3536,6 @@ describe(`MainP${IMPLEMENTATION} contract`, () => {
 
       // Check BU price
       await expectPrice(basketHandler.address, fp('0.75'), ORACLE_ERROR, true)
-    })
-
-    it('lotPrice (deprecated) is equal to price()', async () => {
-      const lotPrice = await basketHandler.lotPrice()
-      const price = await basketHandler.price()
-      expect(price.length).to.equal(2)
-      expect(lotPrice.length).to.equal(price.length)
-      expect(lotPrice[0]).to.equal(price[0])
-      expect(lotPrice[1]).to.equal(price[1])
     })
 
     it('Should not put backup tokens with different targetName in the basket', async () => {