+++ test/integration/AssetPlugins.test.ts
@@ -2425,10 +2425,13 @@ describeFork(`Asset Plugins - Integration - Mainnet Forking P${IMPLEMENTATION}`,
         await expect(rToken.connect(addr1).issue(issueAmount)).to.emit(rToken, 'Issuance')
 
         // Check Balances after
-        expect(await usdt.balanceOf(backingManager.address)).to.equal(toBNDecimals(issueAmount, 6)) //1 full unit
+        const usdtBal = toBNDecimals(issueAmount, 6)
+        expect(await usdt.balanceOf(backingManager.address)).to.be.gt(usdtBal)
+        expect(await usdt.balanceOf(backingManager.address)).to.be.closeTo(usdtBal, 1000) //1 full unit
+
         // Balances for user
         expect(await usdt.balanceOf(addr1.address)).to.equal(
-          toBNDecimals(initialBal.sub(issueAmount), 6)
+          toBNDecimals(initialBal.sub(await usdt.balanceOf(backingManager.address)), 6)
         )
 
         // Check RTokens issued to user