+++ test/plugins/Collateral.test.ts
@@ -397,6 +397,58 @@ describe('Collateral contracts', () => {
       ).to.be.revertedWith('delayUntilDefault too long')
     })
 
+    it('Should not allow low defaultThreshold', async () => {
+      await expect(
+        FiatCollateralFactory.deploy({
+          priceTimeout: PRICE_TIMEOUT,
+          chainlinkFeed: await tokenCollateral.chainlinkFeed(),
+          oracleError: ORACLE_ERROR,
+          erc20: token.address,
+          maxTradeVolume: config.rTokenMaxTradeVolume,
+          oracleTimeout: ORACLE_TIMEOUT,
+          targetName: ethers.utils.formatBytes32String('USD'),
+          defaultThreshold: fp('0.69'),
+          delayUntilDefault: bn(1),
+        })
+      ).to.be.revertedWith('defaultThreshold too low')
+
+      // ATokenFiatCollateral
+      await expect(
+        ATokenFiatCollateralFactory.deploy(
+          {
+            priceTimeout: PRICE_TIMEOUT,
+            chainlinkFeed: await tokenCollateral.chainlinkFeed(),
+            oracleError: ORACLE_ERROR,
+            erc20: aToken.address,
+            maxTradeVolume: config.rTokenMaxTradeVolume,
+            oracleTimeout: ORACLE_TIMEOUT,
+            targetName: ethers.utils.formatBytes32String('USD'),
+            defaultThreshold: fp('0.69'),
+            delayUntilDefault: bn(1),
+          },
+          REVENUE_HIDING
+        )
+      ).to.be.revertedWith('defaultThreshold too low')
+
+      // CTokenFiatCollateral
+      await expect(
+        CTokenFiatCollateralFactory.deploy(
+          {
+            priceTimeout: PRICE_TIMEOUT,
+            chainlinkFeed: await tokenCollateral.chainlinkFeed(),
+            oracleError: ORACLE_ERROR,
+            erc20: cToken.address,
+            maxTradeVolume: config.rTokenMaxTradeVolume,
+            oracleTimeout: ORACLE_TIMEOUT,
+            targetName: ethers.utils.formatBytes32String('USD'),
+            defaultThreshold: fp('0.69'),
+            delayUntilDefault: bn(1),
+          },
+          REVENUE_HIDING
+        )
+      ).to.be.revertedWith('defaultThreshold too low')
+    })
+
     it('Should not allow missing referenceERC20Decimals', async () => {
       // CTokenFiatCollateral with decimals = 0 in underlying
       const token0decimals: BadERC20 = await (